airflow/requirements.txt
apache-airflow==2.8.0
psycopg2-binary==2.9.9
pyspark==3.5.0